Move ByteBufferPrimitiveArray to test support and rearrange test support packages
