+++ message-encoding/src/integration-test/java/com/lmax/solana4j/programs/AssociatedTokenProgramIntegrationTest.java
@@ -1,6 +1,6 @@
 package com.lmax.solana4j.programs;
 
-import com.lmax.solana4j.parameterization.ParameterizedTokenTest;
+import com.lmax.solana4j.parameterisation.ParameterizedTokenTest;
 import org.junit.jupiter.api.BeforeEach;
 
 public class AssociatedTokenProgramIntegrationTest extends SolanaProgramsIntegrationTestBase
