+++ message-encoding/src/integration-test/java/com/lmax/solana4j/programs/BpfLoaderUpgradeableProgramIntegrationTest.java
@@ -1,6 +1,6 @@
 package com.lmax.solana4j.programs;
 
-import com.lmax.solana4j.parameterization.ParameterizedMessageEncodingTest;
+import com.lmax.solana4j.parameterisation.ParameterizedMessageEncodingTest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.parallel.Execution;
