+++ message-encoding/src/integration-test/java/com/lmax/solana4j/programs/SystemProgramIntegrationTest.java
@@ -1,6 +1,6 @@
 package com.lmax.solana4j.programs;
 
-import com.lmax.solana4j.parameterization.ParameterizedMessageEncodingTest;
+import com.lmax.solana4j.parameterisation.ParameterizedMessageEncodingTest;
 import org.junit.jupiter.api.BeforeEach;
 
 public class SystemProgramIntegrationTest extends SolanaProgramsIntegrationTestBase
