+++ message-encoding/src/test-support/java/com/lmax/solana4j/domain/TestKeyPairGenerator.java
@@ -1,6 +1,5 @@
-package com.lmax.solana4j.util;
+package com.lmax.solana4j.domain;
 
-import com.lmax.solana4j.domain.TestKeyPair;
 import org.bouncycastle.math.ec.rfc8032.Ed25519;
 
 import java.security.SecureRandom;
