+++ message-encoding/src/test-support/java/com/lmax/solana4j/transaction/ByteBufferPrimitiveArray.java
@@ -1,0 +1,26 @@
+package com.lmax.solana4j.transaction;
+
+import java.nio.ByteBuffer;
+
+import static java.util.Objects.requireNonNull;
+
+public final class ByteBufferPrimitiveArray
+{
+    private ByteBufferPrimitiveArray()
+    {
+    }
+
+    public static byte[] copy(final ByteBuffer buffer)
+    {
+        requireNonNull(buffer);
+        if (!buffer.hasArray())
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        final var offset = buffer.arrayOffset();
+        final var bytes = new byte[buffer.limit()];
+        System.arraycopy(buffer.array(), offset, bytes, 0, bytes.length);
+        return bytes;
+    }
+}
