+++ message-encoding/src/test-support/java/com/lmax/solana4j/transaction/LegacyTransactionBlobFactory.java
@@ -1,4 +1,4 @@
-package com.lmax.solana4j.transactionblobs;
+package com.lmax.solana4j.transaction;
 
 import com.lmax.solana4j.Solana;
 import com.lmax.solana4j.api.AddressLookupTable;
@@ -17,8 +17,7 @@ import com.lmax.solana4j.programs.ComputeBudgetProgram;
 import com.lmax.solana4j.programs.SystemProgram;
 import com.lmax.solana4j.programs.TokenProgramBase;
 import com.lmax.solana4j.util.Base58;
-import com.lmax.solana4j.util.BouncyCastleSigner;
-import com.lmax.solana4j.util.ByteBufferPrimitiveArray;
+import com.lmax.solana4j.sign.BouncyCastleSigner;
 
 import java.nio.ByteBuffer;
 import java.util.List;
